\name{gnet_bcell}
\alias{gnet}
\docType{data}
\title{An igraph object of B-cell network inferred from gene expression data}
\format{igaph class object}
\usage{
data("gnet_bcell")
}
\description{
A connected subnetwork of bcell network.
}
%\references{de Matos Simoes R, Emmert-Streib F., Bagging statistical network inference from large-scale gene expression data., PLoS One. 2012;7(3):e33624. Epub 2012 Mar 30 }
\keyword{datasets}

