\name{modules_bcell}
\alias{mod.list}
\docType{data}
\title{A list object of modules's information of B-cell network}
\format{list object}
\usage{
data("modules_bcell")
}
\description{
A list object of modules obtained by using fastgreedy algorithm in igraph. Then we performed the enrichment analysis for GO pathways on modules.
}
\keyword{datasets}


