\name{modules_PPI_Athalina}
\alias{lm}
\docType{data}
\title{A list object of modules's information of PPI interaction of A. thalina}
\format{list object}
\usage{
data(modules_PPI_Athalina)
}
\description{
A list object of modules obtained by using fastgreedy algorithm in igraph.
Then we performed the enrichment analysis for GO pathways on modules.
}
\keyword{datasets}


