\name{PPI_Athalina}
\alias{g1}
\docType{data}
\title{An igraph object of PPI interaction of A. thalina}
\format{igaph class object}
\usage{
data(PPI_Athalina)
}
\description{
A connected subnetwork of PPI interaction of A. thalina.
}
\examples{
data("PPI_Athalina")
#xx <- plot.modules(g1)
}
\keyword{datasets}
